+++ Realm/RLMRealm.h
@@ -19,7 +19,7 @@
 #import <Foundation/Foundation.h>
 #import <Realm/RLMDefines.h>
 
-@class RLMObject, RLMSchema, RLMMigration, RLMNotificationToken;
+@class RLMRealmConfiguration, RLMObject, RLMSchema, RLMMigration, RLMNotificationToken;
 
 RLM_ASSUME_NONNULL_BEGIN
 
@@ -59,10 +59,25 @@ RLM_ASSUME_NONNULL_BEGIN
  default Realm is persisted as default.realm under the Documents directory of
  your Application on iOS, and in your application's Application Support
  directory on OS X.
+ 
+ The default Realm is created using the default `RLMRealmConfiguration`, which
+ can be changed via `+[RLMRealmConfiguration setDefaultConfiguration:]`.
 
  @return The default `RLMRealm` instance for the current thread.
  */
 + (instancetype)defaultRealm;
+
+/**
+ Obtains an `RLMRealm` instance with the given configuration.
+
+ @param configuration The configuration for the realm.
+ @param error         If an error occurs, upon return contains an `NSError` object
+                      that describes the problem. If you are not interested in
+                      possible errors, pass in `NULL`.
+
+ @return An `RLMRealm` instance.
+ */
++ (nullable instancetype)realmWithConfiguration:(RLMRealmConfiguration *)configuration error:(NSError **)error;
 
 /**
  Obtains an `RLMRealm` instance persisted at a specific file path.
@@ -94,7 +109,7 @@ RLM_ASSUME_NONNULL_BEGIN
 
  @return An `RLMRealm` instance.
  */
-+ (nullable instancetype)realmWithPath:(NSString *)path readOnly:(BOOL)readonly error:(NSError **)error;
++ (nullable instancetype)realmWithPath:(NSString *)path readOnly:(BOOL)readonly error:(NSError **)error DEPRECATED_MSG_ATTRIBUTE("Use +[RLMRealm realmWithConfiguration:error:]");
 
 /**
  Obtains an `RLMRealm` instance persisted to an encrypted file.
@@ -121,7 +136,7 @@ RLM_ASSUME_NONNULL_BEGIN
 + (nullable instancetype)realmWithPath:(NSString *)path
                          encryptionKey:(NSData *)key
                               readOnly:(BOOL)readonly
-                                 error:(NSError **)error;
+                                 error:(NSError **)error DEPRECATED_MSG_ATTRIBUTE("Use +[RLMRealm realmWithConfiguration:error:]");
 
 /**
  Set the encryption key to use when opening Realms at a certain path.
@@ -138,7 +153,7 @@ RLM_ASSUME_NONNULL_BEGIN
  @param key     64-byte encryption key to use, or `nil` to unset.
  @param path    Realm path to set the encryption key for.
  */
-+ (void)setEncryptionKey:(nullable NSData *)key forRealmsAtPath:(NSString *)path;
++ (void)setEncryptionKey:(nullable NSData *)key forRealmsAtPath:(NSString *)path DEPRECATED_MSG_ATTRIBUTE("Use RLMRealmConfiguration to set an encryption key");
 
 /**
  Obtains an `RLMRealm` instance for an un-persisted in-memory Realm. The identifier
@@ -157,7 +172,7 @@ RLM_ASSUME_NONNULL_BEGIN
 
  @return An `RLMRealm` instance.
  */
-+ (instancetype)inMemoryRealmWithIdentifier:(NSString *)identifier;
++ (instancetype)inMemoryRealmWithIdentifier:(NSString *)identifier DEPRECATED_MSG_ATTRIBUTE("Use +[RLMRealm realmWithConfiguration:error:]");
 
 /**
  Path to the file where this Realm is persisted.
@@ -184,6 +199,11 @@ RLM_ASSUME_NONNULL_BEGIN
  */
 @property (nonatomic, readonly) BOOL inWriteTransaction;
 
+/**
+ Returns an `RLMRealmConfiguration` that can be used to create this `RLMRealm` instance.
+ */
+@property (nonatomic, readonly) RLMRealmConfiguration *configuration;
+
 /**---------------------------------------------------------------------------------------
  *  @name Default Realm Path
  * ---------------------------------------------------------------------------------------
@@ -199,7 +219,7 @@ RLM_ASSUME_NONNULL_BEGIN
 
  @see defaultRealm
  */
-+ (NSString *)defaultRealmPath;
++ (NSString *)defaultRealmPath DEPRECATED_MSG_ATTRIBUTE("Use [RLMRealmConfiguration defaultConfiguration].path");
 
 /**
  Set the default Realm path to a given path.
@@ -208,7 +228,7 @@ RLM_ASSUME_NONNULL_BEGIN
 
  @see defaultRealm
  */
-+ (void)setDefaultRealmPath:(NSString *)defaultRealmPath;
++ (void)setDefaultRealmPath:(NSString *)defaultRealmPath DEPRECATED_MSG_ATTRIBUTE("Use +[RLMRealmConfiguration setDefaultConfiguration:]");
 
 
 #pragma mark - Notifications
@@ -526,7 +546,7 @@ typedef void (^RLMMigrationBlock)(RLMMigration *migration, uint64_t oldSchemaVer
 
  @see               RLMMigration
  */
-+ (void)setDefaultRealmSchemaVersion:(uint64_t)version withMigrationBlock:(nullable RLMMigrationBlock)block;
++ (void)setDefaultRealmSchemaVersion:(uint64_t)version withMigrationBlock:(nullable RLMMigrationBlock)block DEPRECATED_MSG_ATTRIBUTE("Use RLMRealmConfiguration.schemaVersion and RLMRealmConfiguration.migrationBlock");
 
 /**
  Specify a schema version and an associated migration block which is applied when
@@ -538,7 +558,7 @@ typedef void (^RLMMigrationBlock)(RLMMigration *migration, uint64_t oldSchemaVer
 
  @see               RLMMigration
  */
-+ (void)setSchemaVersion:(uint64_t)version forRealmAtPath:(NSString *)realmPath withMigrationBlock:(nullable RLMMigrationBlock)block;
++ (void)setSchemaVersion:(uint64_t)version forRealmAtPath:(NSString *)realmPath withMigrationBlock:(nullable RLMMigrationBlock)block DEPRECATED_MSG_ATTRIBUTE("Use RLMRealmConfiguration.schemaVersion and RLMRealmConfiguration.migrationBlock");
 
 /**
  Get the schema version for a Realm at a given path.
@@ -566,6 +586,20 @@ typedef void (^RLMMigrationBlock)(RLMMigration *migration, uint64_t oldSchemaVer
 + (uint64_t)schemaVersionAtPath:(NSString *)realmPath encryptionKey:(nullable NSData *)key error:(NSError **)error;
 
 /**
+ Performs the given Realm configuration's migration block on a Realm at the given path.
+
+ This method is called automatically when opening a Realm for the first time and does
+ not need to be called explicitly. You can choose to call this method to control
+ exactly when and how migrations are performed.
+
+ @param configuration The Realm configuration used to open and migrate the Realm.
+ @return              The error that occurred while applying the migration, if any.
+
+ @see                 RLMMigration
+ */
++ (NSError *)migrateRealm:(RLMRealmConfiguration *)configuration;
+
+/**
  Performs the registered migration block on a Realm at the given path.
 
  This method is called automatically when opening a Realm for the first time and does
@@ -578,7 +612,7 @@ typedef void (^RLMMigrationBlock)(RLMMigration *migration, uint64_t oldSchemaVer
  @see               RLMMigration
  @see               setSchemaVersion:forRealmAtPath:withMigrationBlock:
  */
-+ (NSError *)migrateRealmAtPath:(NSString *)realmPath;
++ (NSError *)migrateRealmAtPath:(NSString *)realmPath DEPRECATED_MSG_ATTRIBUTE("Use +[RLMRealm migrateRealm:]");
 
 /**
  Performs the registered migration block on an encrypted Realm at the given path.
@@ -589,7 +623,7 @@ typedef void (^RLMMigrationBlock)(RLMMigration *migration, uint64_t oldSchemaVer
  @param key         64-byte encryption key.
  @return            The error that occurred while applying the migration, if any.
  */
-+ (NSError *)migrateRealmAtPath:(NSString *)realmPath encryptionKey:(NSData *)key;
++ (NSError *)migrateRealmAtPath:(NSString *)realmPath encryptionKey:(NSData *)key DEPRECATED_MSG_ATTRIBUTE("Use +[RLMRealm migrateRealm:]");
 
 #pragma mark -