+++ Realm/RLMRealm_Private.h
@@ -23,6 +23,14 @@
 // Disable syncing files to disk. Cannot be re-enabled. Use only for tests.
 FOUNDATION_EXTERN void RLMDisableSyncToDisk();
 
+FOUNDATION_EXTERN NSData *RLMRealmValidatedEncryptionKey(NSData *key);
+
+FOUNDATION_EXTERN void RLMRealmSetEncryptionKeyForPath(NSData *encryptionKey, NSString *path);
+
+FOUNDATION_EXTERN void RLMRealmSetSchemaVersionForPath(uint64_t version, NSString *path, RLMMigrationBlock migrationBlock);
+
+FOUNDATION_EXTERN void RLMRealmAddPathSettingsToConfiguration(RLMRealmConfiguration *configuration);
+
 // RLMRealm private members
 @interface RLMRealm () {
     @public