+++ RealmSwift-swift1.2/Realm.swift
@@ -55,6 +55,9 @@ public final class Realm {
     /// The Schema used by this realm.
     public var schema: Schema { return Schema(rlmRealm.schema) }
 
+    /// Returns a `Configuration` that can be used to create this `Realm` instance.
+    public var configuration: Configuration { return Configuration.fromRLMRealmConfiguration(rlmRealm.configuration) }
+
     /**
     The location of the default Realm as a string. Can be overridden.
 
@@ -64,27 +67,53 @@ public final class Realm {
 
     :returns: Location of the default Realm.
     */
+    @availability(*, deprecated=1, message="Use Configuration.defaultConfiguration")
     public class var defaultPath: String {
         get {
-            return RLMRealm.defaultRealmPath()
+            return Configuration.defaultConfiguration.path ?? RLMRealmConfiguration.defaultRealmPath()
         }
         set {
-            RLMRealm.setDefaultRealmPath(newValue)
+            RLMRealmConfiguration.setDefaultPath(newValue)
         }
     }
 
     // MARK: Initializers
 
     /**
-    Obtains a Realm instance persisted at the specified file path. Defaults to
-    `Realm.defaultPath`
+    Obtains a Realm instance with the given configuration. Defaults to the default Realm configuration,
+    which can be changed by setting `Realm.Configuration.defaultConfiguration`.
+
+    :param: configuration The configuration to use when creating the Realm instance.
+    :param: error         If an error occurs, upon return contains an `NSError` object
+                          that describes the problem. If you are not interested in
+                          possible errors, omit the argument, or pass in `nil`.
+    */
+    public convenience init?(configuration: Configuration, error: NSErrorPointer = nil)  {
+        if let rlmRealm = RLMRealm(configuration: configuration.rlmConfiguration, error: error) {
+            self.init(rlmRealm)
+        } else {
+            self.init(RLMRealm())
+            return nil
+        }
+    }
+
+    /**
+    Obtains a Realm instance with the default `Realm.Configuration`.
+    */
+    public convenience init() {
+        let rlmRealm = RLMRealm.defaultRealm()
+        self.init(rlmRealm)
+    }
+
+    /**
+    Obtains a Realm instance persisted at the specified file path.
 
     :param: path Path to the realm file.
     */
-    public convenience init(path: String = Realm.defaultPath) {
-        self.init(RLMRealm(path: path, readOnly: false, error: nil)!)
+    public convenience init(path: String) {
+        self.init(RLMRealm(path: path, key: nil, readOnly: false, inMemory: false, dynamic: false, schema: nil, error: nil)!)
     }
-    
+
     /**
     Obtains a `Realm` instance with persistence to a specific file path with
     options.
@@ -105,6 +134,7 @@ public final class Realm {
                             that describes the problem. If you are not interested in
                             possible errors, omit the argument, or pass in `nil`.
     */
+    @availability(*, deprecated=1, message="Use Realm(configuration:error:)")
     public convenience init?(path: String, readOnly: Bool, encryptionKey: NSData? = nil, error: NSErrorPointer = nil) {
         if let rlmRealm = RLMRealm(path: path, key: encryptionKey, readOnly: readOnly, inMemory: false, dynamic: false, schema: nil, error: error) as RLMRealm? {
             self.init(rlmRealm)
@@ -129,8 +159,10 @@ public final class Realm {
 
     :param: identifier A string used to identify a particular in-memory Realm.
     */
+    @availability(*, deprecated=1, message="Use Realm(configuration:error:)")
     public convenience init(inMemoryIdentifier: String) {
-        self.init(RLMRealm.inMemoryRealmWithIdentifier(inMemoryIdentifier))
+        let configuration = Configuration(inMemoryIdentifier: inMemoryIdentifier).rlmConfiguration
+        self.init(RLMRealm(configuration: configuration, error: nil)!)
     }
 
     // MARK: Transactions
@@ -268,9 +300,9 @@ public final class Realm {
 
     Creates or updates an instance of this object and adds it to the `Realm` populating
     the object with the given value.
-    
+
     When 'update' is 'true', the object must have a primary key. If no objects exist in
-    the Realm instance with the same primary key value, the object is inserted. Otherwise, 
+    the Realm instance with the same primary key value, the object is inserted. Otherwise,
     the existing object is updated with any changed values.
 
     :param: type    The object type to create.
@@ -320,7 +352,7 @@ public final class Realm {
 
     :param: object The objects to be deleted. This can be a `List<Object>`, `Results<Object>`,
                    or any other enumerable SequenceType which generates Object.
-    
+
     :nodoc:
     */
     public func delete<T: Object>(objects: List<T>) {
@@ -332,7 +364,7 @@ public final class Realm {
 
     :param: object The objects to be deleted. This can be a `List<Object>`, `Results<Object>`,
                    or any other enumerable SequenceType which generates Object.
-    
+
     :nodoc:
     */
     public func delete<T: Object>(objects: Results<T>) {
@@ -525,8 +557,9 @@ public final class Realm {
     :param: encryptionKey 64-byte encryption key to use, or `nil` to unset.
     :param: path          Realm path to set the encryption key for.
     +*/
-    public class func setEncryptionKey(encryptionKey: NSData?, forPath: String = Realm.defaultPath) {
-        RLMRealm.setEncryptionKey(encryptionKey, forRealmsAtPath: forPath)
+    @availability(*, deprecated=1, message="Use Realm(configuration:error:)")
+    public class func setEncryptionKey(encryptionKey: NSData?, forPath path: String = Realm.defaultPath) {
+        RLMRealmSetEncryptionKeyForPath(encryptionKey, path)
     }
 
     // MARK: Internal