+++ RealmSwift-swift2.0/Realm.swift
@@ -55,6 +55,9 @@ public final class Realm {
     /// The Schema used by this realm.
     public var schema: Schema { return Schema(rlmRealm.schema) }
 
+    /// Returns a `Configuration` that can be used to create this `Realm` instance.
+    public var configuration: Configuration { return Configuration.fromRLMRealmConfiguration(rlmRealm.configuration) }
+
     /**
     The location of the default Realm as a string. Can be overridden.
 
@@ -64,28 +67,41 @@ public final class Realm {
 
     - returns: Location of the default Realm.
     */
+    @available(*, deprecated=1, message="Use Realm.Configuration.defaultConfiguration")
     public class var defaultPath: String {
         get {
-            return RLMRealm.defaultRealmPath()
+            return Configuration.defaultConfiguration.path ?? RLMRealmConfiguration.defaultRealmPath()
         }
         set {
-            RLMRealm.setDefaultRealmPath(newValue)
+            RLMRealmConfiguration.setDefaultPath(newValue)
         }
     }
 
     // MARK: Initializers
 
     /**
-    Obtains a Realm instance persisted at the specified file path. Defaults to
-    `Realm.defaultPath`
+    Obtains a Realm instance with the given configuration. Defaults to the default Realm configuration,
+    which can be changed by setting `Realm.Configuration.defaultConfiguration`.
+
+    - parameter configuration: The configuration to use when creating the Realm instance.
+    */
+    public convenience init(configuration: Configuration = Configuration.defaultConfiguration) throws {
+        let rlmConfiguration = configuration.rlmConfiguration
+        RLMRealmAddPathSettingsToConfiguration(rlmConfiguration)
+        let rlmRealm = try RLMRealm(configuration: rlmConfiguration)
+        self.init(rlmRealm)
+    }
+
+    /**
+    Obtains a Realm instance persisted at the specified file path.
 
     - parameter path: Path to the realm file.
     */
-    public convenience init(path: String = Realm.defaultPath) throws {
+    public convenience init(path: String) throws {
         let rlmRealm = try RLMRealm(path: path, key: nil, readOnly: false, inMemory: false, dynamic: false, schema: nil)
         self.init(rlmRealm)
     }
-    
+
     /**
     Obtains a `Realm` instance with persistence to a specific file path with
     options.
@@ -103,6 +119,7 @@ public final class Realm {
     - parameter readOnly:      Bool indicating if this Realm is read-only (must use for read-only files).
     - parameter encryptionKey: 64-byte key to use to encrypt the data.
     */
+    @available(*, deprecated=1, message="Use Realm(configuration:)")
     public convenience init(path: String, readOnly: Bool, encryptionKey: NSData? = nil) throws {
         let rlmRealm = try RLMRealm(path: path, key: encryptionKey, readOnly: readOnly, inMemory: false, dynamic: false, schema: nil)
         self.init(rlmRealm)
@@ -123,8 +140,10 @@ public final class Realm {
 
     - parameter identifier: A string used to identify a particular in-memory Realm.
     */
-    public convenience init(inMemoryIdentifier: String) {
-        self.init(RLMRealm.inMemoryRealmWithIdentifier(inMemoryIdentifier))
+    @available(*, deprecated=1, message="Use Realm(configuration:)")
+    public convenience init(inMemoryIdentifier: String) throws {
+        let configuration = Configuration(inMemoryIdentifier: inMemoryIdentifier)
+        try self.init(configuration: configuration)
     }
 
     // MARK: Transactions
@@ -262,9 +281,9 @@ public final class Realm {
 
     Creates or updates an instance of this object and adds it to the `Realm` populating
     the object with the given value.
-    
+
     When 'update' is 'true', the object must have a primary key. If no objects exist in
-    the Realm instance with the same primary key value, the object is inserted. Otherwise, 
+    the Realm instance with the same primary key value, the object is inserted. Otherwise,
     the existing object is updated with any changed values.
 
     - parameter type:   The object type to create.
@@ -315,7 +334,7 @@ public final class Realm {
 
     - parameter object: The objects to be deleted. This can be a `List<Object>`, `Results<Object>`,
                         or any other enumerable SequenceType which generates Object.
-    
+
     :nodoc:
     */
     public func delete<T: Object>(objects: List<T>) {
@@ -327,7 +346,7 @@ public final class Realm {
 
     - parameter object: The objects to be deleted. This can be a `List<Object>`, `Results<Object>`,
                         or any other enumerable SequenceType which generates Object.
-    
+
     :nodoc:
     */
     public func delete<T: Object>(objects: Results<T>) {
@@ -515,9 +534,10 @@ public final class Realm {
 
     - parameter encryptionKey: 64-byte encryption key to use, or `nil` to unset.
     - parameter path:          Realm path to set the encryption key for.
-    +*/
-    public class func setEncryptionKey(encryptionKey: NSData?, forPath: String = Realm.defaultPath) {
-        RLMRealm.setEncryptionKey(encryptionKey, forRealmsAtPath: forPath)
+    */
+    @available(*, deprecated=1, message="Use Realm(configuration:)")
+    public class func setEncryptionKey(encryptionKey: NSData?, forPath path: String = Realm.defaultPath) {
+        RLMRealmSetEncryptionKeyForPath(encryptionKey, path)
     }
 
     // MARK: Internal